Pull truncatenumber out of Lfunctionsutilities
Basic number utilities should be available at a basic directory, and
subdirectories should adjust them as necessary. This commit moves
truncatenumber out of lmfdb.lfunctions.Lfunctionsutilities and into
lmfdb.utils. By doing this, it is now possible to import truncatenumber
into mwf_classes.py at the module level instead of at the function
level, and a circular import has been avoided.

Additional utilities refactoring should take place, but I defer that to
another issue and pull request. This led to #2128.
