+++ lmfdb/utils.py
@@ -640,4 +640,27 @@ def rgbtohex(rgb):
 
     return "#{:02x}{:02x}{:02x}".format(r,g,b)
 
-
+def truncatenumber(numb, precision):
+    localprecision = precision
+    if numb < 0:
+        localprecision = localprecision + 1
+    truncation = float(10 ** (-1.0*localprecision))
+    if float(abs(numb - 1)) < truncation:
+        return("1")
+    elif float(abs(numb - 2)) < truncation:
+        return("2")
+    elif float(abs(numb - 3)) < truncation:
+        return("3")
+    elif float(abs(numb - 4)) < truncation:
+        return("4")
+    elif float(abs(numb)) < truncation:
+        return("0")
+    elif float(abs(numb + 1)) < truncation:
+        return("-1")
+    elif float(abs(numb + 2)) < truncation:
+        return("-2")
+    elif float(abs(numb - 0.5)) < truncation:
+        return("0.5")
+    elif float(abs(numb + 0.5)) < truncation:
+        return("-0.5")
+    return(str(numb)[0:int(localprecision)])
